feat(roxy-cli): CLI 및 명령어 모듈 추가 및 구현
