+++ roxy-cli/src/lib.rs
@@ -1,0 +1,5 @@
+// 라이브러리로 노출할 모듈들
+
+pub mod cli;
+pub mod commands;
+pub mod docker;
